refactor: moved view to core ui
