+++ app/src/main/java/rees46/demo_android/core_ui/RecommendationBlockView.kt
@@ -1,4 +1,4 @@
-package rees46.demo_android.features.recommendationBlock
+package rees46.demo_android.core_ui
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -8,6 +8,7 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.RecyclerView
 import rees46.demo_android.R
 import rees46.demo_android.features.product.Product
+import rees46.demo_android.features.recommendationBlock.ShortCardProductsAdapter
 
 class RecommendationBlockView @JvmOverloads constructor(
     context: Context, attrs: AttributeSet? = null
