+++ app/src/main/java/rees46/demo_android/features/main/home/HomeFragment.kt
@@ -4,11 +4,9 @@ import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import rees46.demo_android.R
 import rees46.demo_android.core.view.BaseFragment
 import rees46.demo_android.databinding.FragmentHomeBinding
-import rees46.demo_android.features.product.ModelBuilder
-import rees46.demo_android.features.recommendationBlock.RecommendationBlockView
+import rees46.demo_android.core_ui.RecommendationBlockView
 
 class HomeFragment
     : BaseFragment<FragmentHomeBinding>(FragmentHomeBinding::inflate) {
