+++ app/src/main/java/rees46/demo_android/features/recommendationBlock/ShortCardProductsAdapter.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import rees46.demo_android.core_ui.ShortCardProductView
 import rees46.demo_android.features.product.Product
 
 class ShortCardProductsAdapter(
