+++ app/src/main/res/layout/fragment_card_product.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_width="match_parent">
 
-    <rees46.demo_android.features.cardProduct.CardProductView
+    <rees46.demo_android.core_ui.CardProductView
         android:id="@+id/cardProductView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -13,7 +13,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <rees46.demo_android.features.recommendationBlock.RecommendationBlockView
+    <rees46.demo_android.core_ui.RecommendationBlockView
         android:id="@+id/recommendation_block"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
